+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -120,6 +120,7 @@
     <Compile Include="Memory\VmdPrototype.cs" />
     <Compile Include="IRealTime.cs" />
     <Compile Include="ListFilters.cs" />
+    <Compile Include="LogConsole.cs" />
     <Compile Include="MemoryBanks.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="BlastGeneratorProto.cs" />
